Refactor file handler to use separate parser and writer
