+++ repo_agent/file_handler.py
@@ -1,5 +1,3 @@
-import ast
-import json
 import os
 
 import git
@@ -10,6 +8,8 @@ from repo_agent.log import logger
 from repo_agent.settings import SettingsManager
 from repo_agent.utils.gitignore_checker import GitignoreChecker
 from repo_agent.utils.meta_info_utils import latest_version_substring
+from repo_agent.ast_parser import ASTParser
+from repo_agent.markdown_writer import MarkdownWriter
 
 
 class FileHandler:
@@ -32,6 +32,9 @@ class FileHandler:
             setting.project.target_repo / setting.project.hierarchy_name
         )
 
+        self.ast_parser = ASTParser(repo_path)
+        self.markdown_writer = MarkdownWriter()
+
     def _read_file_content(self, relative_path):
         """
         Safely read file content given a path relative to the repo root.
@@ -51,53 +54,6 @@ class FileHandler:
         """
         return self._read_file_content(self.file_path)
 
-    def get_obj_code_info(
-        self, code_type, code_name, start_line, end_line, params, file_path=None
-    ):
-        """
-        Get the code information for a given object (function/class/async).
-
-        Args:
-            code_type (str): The type of the code ('FunctionDef', 'ClassDef', etc.).
-            code_name (str): The name of the code object.
-            start_line (int): The starting line number of the code.
-            end_line (int): The ending line number of the code.
-            params (list): List of parameter names.
-            file_path (str, optional): Relative file path. Defaults to None.
-
-        Returns:
-            dict: A dictionary containing code details (type, name, content, etc.).
-        """
-
-        code_info = {}
-        code_info["type"] = code_type
-        code_info["name"] = code_name
-        code_info["md_content"] = []
-        code_info["code_start_line"] = start_line
-        code_info["code_end_line"] = end_line
-        code_info["params"] = params
-
-        target_path = file_path if file_path is not None else self.file_path
-        full_path = os.path.join(self.repo_path, target_path)
-
-        with open(full_path, "r", encoding="utf-8") as f:
-            lines = f.readlines()
-            code_content = "".join(lines[start_line - 1:end_line])
-
-            # Find the object's name position in the first line
-            name_column = lines[start_line - 1].find(code_name)
-
-            # Check if there's a "return" in the code block
-            if "return" in code_content:
-                has_return = True
-            else:
-                has_return = False
-
-            code_info["has_return"] = has_return
-            code_info["code_content"] = code_content
-            code_info["name_column"] = name_column
-
-        return code_info
 
     def write_file(self, file_path, content):
         """
@@ -144,84 +100,9 @@ class FileHandler:
 
         return current_version, previous_version
 
-    def get_end_lineno(self, node):
-        """
-        Recursively find the maximum end_lineno among a node and its children.
-
-        Args:
-            node (ast.AST): The AST node.
-
-        Returns:
-            int: The final end line number. -1 if not applicable.
-        """
-        if not hasattr(node, "lineno"):
-            return -1
-
-        end_lineno = node.lineno
-        for child in ast.iter_child_nodes(node):
-            child_end = getattr(child, "end_lineno", None) or self.get_end_lineno(child)
-            if child_end > -1:
-                end_lineno = max(end_lineno, child_end)
-        return end_lineno
-
-    def add_parent_references(self, node, parent=None):
-        """
-        Recursively add 'parent' references to each child node in the AST.
-
-        Args:
-            node (ast.AST): The current AST node.
-            parent (ast.AST, optional): The parent node. Defaults to None.
-        """
-        for child in ast.iter_child_nodes(node):
-            child.parent = node
-            self.add_parent_references(child, node)
-
-    def get_functions_and_classes(self, code_content):
-        """
-        Retrieves functions/classes/async functions from the AST.
-
-        Returns:
-            list: A list of tuples (type, name, start_line, end_line, params).
-        """
-        tree = ast.parse(code_content)
-        self.add_parent_references(tree)
-
-        functions_and_classes = []
-        for node in ast.walk(tree):
-            if isinstance(node, (ast.FunctionDef, ast.ClassDef, ast.AsyncFunctionDef)):
-                start_line = node.lineno
-                end_line = self.get_end_lineno(node)
-                parameters = (
-                    [arg.arg for arg in node.args.args] if "args" in dir(node) else []
-                )
-                functions_and_classes.append(
-                    (type(node).__name__, node.name, start_line, end_line, parameters)
-                )
-
-        return functions_and_classes
-
     def generate_file_structure(self, file_path):
-        """
-        Parse a file via AST and collect code objects' info.
-
-        Args:
-            file_path (str): Relative path of the file.
-
-        Returns:
-            list: A list of dicts with info about each found code object.
-        """
-        content = self._read_file_content(file_path)
-        structures = self.get_functions_and_classes(content)
-
-        file_objects = []
-        for struct in structures:
-            structure_type, name, start_line, end_line, params = struct
-            code_info = self.get_obj_code_info(
-                structure_type, name, start_line, end_line, params, file_path
-            )
-            file_objects.append(code_info)
-
-        return file_objects
+        """Parse a file and collect code objects using :class:`ASTParser`."""
+        return self.ast_parser.generate_file_structure(file_path)
 
     def generate_overall_structure(self, file_path_reflections, jump_files) -> dict:
         """
@@ -271,58 +152,6 @@ class FileHandler:
         return repo_structure
 
     def convert_to_markdown_file(self, file_path=None):
-        """
-        Convert the file structure (as stored in project_hierarchy.json) to a markdown outline.
-
-        Args:
-            file_path (str, optional): The relative path of the file to be converted.
-
-        Returns:
-            str: The content in markdown format.
-
-        Raises:
-            ValueError: If no corresponding entry is found in project_hierarchy.json.
-        """
-        with open(self.project_hierarchy, "r", encoding="utf-8") as f:
-            json_data = json.load(f)
-
-        if file_path is None:
-            file_path = self.file_path
-
-        file_dict = json_data.get(file_path)
-        if file_dict is None:
-            raise ValueError(
-                f"No file object found for {self.file_path} in project_hierarchy.json"
-            )
-
-        markdown = ""
-        parent_dict = {}
-        objects = sorted(file_dict.values(), key=lambda obj: obj["code_start_line"])
-
-        for obj in objects:
-            if obj["parent"] is not None:
-                parent_dict[obj["name"]] = obj["parent"]
-
-        current_parent = None
-        for obj in objects:
-            level = 1
-            parent = obj["parent"]
-            while parent is not None:
-                level += 1
-                parent = parent_dict.get(parent)
-
-            if level == 1 and current_parent is not None:
-                markdown += "***\n"
-            current_parent = obj["name"]
-
-            params_str = ""
-            if obj["type"] in ["FunctionDef", "AsyncFunctionDef"]:
-                params_str = "()"
-                if obj["params"]:
-                    params_str = f"({', '.join(obj['params'])})"
-
-            markdown += f"{'#' * level} {obj['type']} {obj['name']}{params_str}:\n"
-            markdown += f"{obj['md_content'][-1] if len(obj['md_content']) > 0 else ''}\n"
-
-        markdown += "***\n"
-        return markdown
+        """Return Markdown outline for ``file_path`` using :class:`MarkdownWriter`."""
+        target = file_path or self.file_path
+        return self.markdown_writer.convert_to_markdown_file(target)
